+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   12 weeks. Introduced a preliminary filtering by the QC state, which is
   now available in ml warehouse. Since the ml warehouse QC state might not
   be up-to-date, a final check against the LangQC database is retained.
+* Major upgrade of FastAPI, Pydantic and related dependencies.
 
 ## [1.4.1] - 2023-08-23